Remove environment variables that start with dashes (-)
A nice option for option parsing in Win32 is https://stackoverflow.com/a/8162578 which uses environment variables that start with dashes to process command line options. However, if these are present in the environment then when written in scripts (ex. with-compiler.sh) we get command lines like `/usr/bin/env -option1 value1 -option2 value2 COMSPEC=... DKML_TARGET_ABI=...`. So `-option1=value1` in the environment will cause `-option1` to be an argument for /usr/bin/env, which will fail or escape in bad ways.